fix uefi?

+++ node/create-iso.sh
@@ -8,6 +8,12 @@ then
     exit 1
 fi
 
+if [ ! -f /usr/lib/syslinux/bios/isohdpfx.bin ];
+then
+    echo "In order to run this script you need to have the 'syslinux' binary files on your machine."
+    exit 1
+fi
+
 curl https://download.rockylinux.org/pub/rocky/9/isos/x86_64/Rocky-9.3-x86_64-minimal.iso -C - -o rocky.iso
 
 mkdir -p rocky-iso
@@ -15,7 +21,7 @@ mkdir -p rocky-iso
     cd rocky-iso
     bsdtar xf ../rocky.iso
     cd ..
-    cp ks.cfg rocky-iso/isolinux/
+    cp ks.cfg rocky-iso/
     cp ks.cfg rocky-iso/ks-efi.cfg
     cp grub.cfg rocky-iso/EFI/BOOT/grub.cfg
     cp ninux-splash.png rocky-iso/isolinux/splash.png
@@ -26,6 +32,7 @@ mkdir -p rocky-iso
         -isohybrid-mbr /usr/lib/syslinux/bios/isohdpfx.bin \
      	-o ../ninux.iso \
         -b isolinux/isolinux.bin \
+        -J -joliet-long \
         -c boot.cat \
         -no-emul-boot -boot-load-size 4 -boot-info-table \
         -eltorito-alt-boot \
@@ -33,7 +40,7 @@ mkdir -p rocky-iso
         -e images/efiboot.img \
         -no-emul-boot \
 	    -isohybrid-gpt-basdat \
-        -V "Rocky-NInux-9" -R -J -v .
+        -V "Rocky-NInux-9" -R -v .
 )
 
 rm -rf rocky-iso
